requirements.txt
py-cord==2.7.1
python-dotenv==1.0.0
Django==5.2.8
django-markdownx==4.0.9
asgiref==3.10.0
humanize==4.15.0
pycryptodome==3.17
qrcode[PIL]==8.2
legacy-cgi; python_version >= '3.13'
audioop-lts; python_version>='3.13'